Fix torch.compile crash: numpy scalar in regularizer coeffs + hidden SIREN consistency
- regularizer.py: wrap anneal() return in float() so np.exp never leaks numpy.float64
  into _coeffs — caused torch.compile(fullgraph=True) to fail on coefficient comparisons.
  Bug was masked when regul_annealing_rate=0.0 (no numpy involved in that branch).

- config.py: move hidden_neuron_fraction to GraphModelConfig so NNR_hidden can be
  built in __init__ with correct output size (consistent with NNR_f for visual SIREN)

- neural_gnn.py: build NNR_hidden in __init__ like NNR_f — no deferred setup_hidden_siren,
  no post-init model mutation, stable structure for torch.compile

- graph_trainer.py / graph_tester.py: read hidden_neuron_fraction from model_config,
  add has_hidden_neurons flag, remove setup_hidden_siren + optimizer rebuild block

- plot.py: apply visible-edge mask only when zero-silencing (not when SIREN active)

- GNN_Main.py: auto-derive config_file from YAML filename when still "none",
  preventing all absolute-path runs from logging to log/none/

- hidden_005/010/020 YAMLs: move hidden_neuron_fraction to graph_model:, add
  config_file field, torch_compile: true, alternate_training: false

+++ src/connectome_gnn/models/regularizer.py
@@ -117,7 +117,7 @@ class LossRegularizer:
         # rate > 0: coeff ramps up over epochs (zero at epoch 0)
         # rate = 0: full coeff from epoch 0 (no annealing)
         def anneal(coeff):
-            return coeff * (1 - np.exp(-rate * epoch)) if rate > 0 else coeff
+            return float(coeff * (1 - np.exp(-rate * epoch))) if rate > 0 else float(coeff)
 
         self._coeffs['W_L1'] = anneal(tc.coeff_W_L1)
         self._coeffs['W_L2'] = anneal(tc.coeff_W_L2)
